Safari-611.1.10/LayoutTests/fast/canvas/canvas-createPattern-fillRect-shadow-expected.txt

Ensure correct behavior of canvas with createPattern + fillRect with shadow.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 51
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 51
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 51
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 141
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 113
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 51
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 29
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 23
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 29
